+++ src/tasks/task-status.enum.ts
@@ -1,0 +1,5 @@
+export enum TaskStatus {
+    OPEN = 'OPEN',
+    IN_PROGRESS = 'IN_PROGRESS',
+    DONE = 'DONE',
+}
