+++ src/tasks/task.entity.ts
@@ -1,0 +1,17 @@
+import { BaseEntity, Column, Entity, PrimaryGeneratedColumn } from "typeorm";
+import { TaskStatus } from './task-status.enum';
+
+@Entity()
+export class Task extends BaseEntity {
+    @PrimaryGeneratedColumn()
+    id: number;
+
+    @Column() 
+    title: string;
+
+    @Column()
+    description: string;
+
+    @Column()
+    status: TaskStatus
+}
