+++ src/tasks/task.repository.ts
@@ -1,0 +1,7 @@
+import { Entity, EntityRepository, Repository } from "typeorm";
+import { Task } from "./task.entity";
+
+@EntityRepository(Task)
+export class TaskRepository extends Repository<Task> {
+
+}
