+++ src/tasks/tasks.controller.ts
@@ -1,6 +1,5 @@
 import { Body, Controller, Delete, Get, Param, Patch, Post, Query, UsePipes, ValidationPipe } from '@nestjs/common';
 import { TasksService } from './tasks.service';
-import { Task, TaskStatus } from './tasks.model';
 import { Deserializer } from 'v8';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
@@ -10,37 +9,37 @@ import { TaskStatusValidationPipe } from './pipes/task-status-validation.pipe';
 export class TasksController {
     constructor( private tasksService: TasksService){}
 
-    @Get()
-    getTasks(@Query(ValidationPipe) filterDto: GetTasksFilterDto): Task[] {
-        if(Object.keys(filterDto).length){
-            return this.tasksService.getTasksWithFilters(filterDto);
-        }else{
-            return this.tasksService.getAllTasks();
-        }
-    }
+    // @Get()
+    // getTasks(@Query(ValidationPipe) filterDto: GetTasksFilterDto): Task[] {
+    //     if(Object.keys(filterDto).length){
+    //         return this.tasksService.getTasksWithFilters(filterDto);
+    //     }else{
+    //         return this.tasksService.getAllTasks();
+    //     }
+    // }
 
-    @Get('/:id')
-    getTaskById(@Param('id') id: string){
-        return this.tasksService.getTaskById(id);
-    }
+    // @Get('/:id')
+    // getTaskById(@Param('id') id: string){
+    //     return this.tasksService.getTaskById(id);
+    // }
 
-    @Post()
-    @UsePipes(ValidationPipe)
-    createTask(@Body() createTaskDto: CreateTaskDto): Task {
-        return this.tasksService.createTask(createTaskDto);
-    }
+    // @Post()
+    // @UsePipes(ValidationPipe)
+    // createTask(@Body() createTaskDto: CreateTaskDto): Task {
+    //     return this.tasksService.createTask(createTaskDto);
+    // }
 
-    @Delete('/:id')
-    deleteTask(@Param('id') id:string): void{
-        this.tasksService.deleteTask(id);
-    }
+    // @Delete('/:id')
+    // deleteTask(@Param('id') id:string): void{
+    //     this.tasksService.deleteTask(id);
+    // }
 
-    @Patch('/:id/status')
-    updateTaskStatus(
-        @Param('id') id: string,
-        @Body('status', TaskStatusValidationPipe) status: TaskStatus,
-    ): Task {
-        return this.tasksService.updateTaskStatus(id, status);
-    }
+    // @Patch('/:id/status')
+    // updateTaskStatus(
+    //     @Param('id') id: string,
+    //     @Body('status', TaskStatusValidationPipe) status: TaskStatus,
+    // ): Task {
+    //     return this.tasksService.updateTaskStatus(id, status);
+    // }
 
 }
