+++ src/tasks/tasks.module.ts
@@ -1,8 +1,13 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { TasksController } from './tasks.controller';
 import { TasksService } from './tasks.service';
+import { TaskRepository } from './task.repository';
 
 @Module({
+  imports: [
+    TypeOrmModule.forFeature([TaskRepository]),
+  ],
   controllers: [TasksController],
   providers: [TasksService]
 })
