+++ src/tasks/tasks.service.ts
@@ -1,70 +1,67 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { Task, TaskStatus } from './tasks.model';
-import {v4 as uuid} from 'uuid';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 
 @Injectable()
 export class TasksService {
 
-    private tasks: Task[] = [];
-    tasksService: any;
+    // tasksService: any;
 
-    getAllTasks(): Task[] {
-        return this.tasks;
-    }
+    // getAllTasks(): Task[] {
+    //     return this.tasks;
+    // }
 
-    getTasksWithFilters(filterDto: GetTasksFilterDto): Task[] {
-        const { status, search } = filterDto;
+    // getTasksWithFilters(filterDto: GetTasksFilterDto): Task[] {
+    //     const { status, search } = filterDto;
 
-        let tasks = this.getAllTasks();
+    //     let tasks = this.getAllTasks();
 
-        if (status) {
-            tasks = tasks.filter(task => task.status === status)
-        }
+    //     if (status) {
+    //         tasks = tasks.filter(task => task.status === status)
+    //     }
 
-        if (search) {
-            tasks = tasks.filter(task => 
-                task.title.includes(search) ||
-                task.description.includes(search),
-            );
-        }
+    //     if (search) {
+    //         tasks = tasks.filter(task => 
+    //             task.title.includes(search) ||
+    //             task.description.includes(search),
+    //         );
+    //     }
 
-        return tasks;
-    }
+    //     return tasks;
+    // }
 
-    getTaskById(id:string): Task {
-        const found = this.tasks.find(task => task.id === id);
-        if(!found) {
-            throw new NotFoundException(`Task with ID "${id}" not found`);
-        }
+    // getTaskById(id:string): Task {
+    //     const found = this.tasks.find(task => task.id === id);
+    //     if(!found) {
+    //         throw new NotFoundException(`Task with ID "${id}" not found`);
+    //     }
 
-        return found;
-    }
+    //     return found;
+    // }
 
 
-    createTask(createTaskDto: CreateTaskDto): Task{
-        const { title,description } = createTaskDto;
-        const task: Task = {
-            id: uuid(),
-            title: title,
-            description: description,
-            status: TaskStatus.OPEN,
-        };
+    // createTask(createTaskDto: CreateTaskDto): Task{
+    //     const { title,description } = createTaskDto;
+    //     const task: Task = {
+    //         id: uuid(),
+    //         title: title,
+    //         description: description,
+    //         status: TaskStatus.OPEN,
+    //     };
 
-        this.tasks.push(task);
-        return task;
-    }
+    //     this.tasks.push(task);
+    //     return task;
+    // }
 
-    deleteTask(id: string): void{
-        const found = this.getTaskById(id);
-        this.tasks = this.tasks.filter(task => task.id !== found.id)
-    }
+    // deleteTask(id: string): void{
+    //     const found = this.getTaskById(id);
+    //     this.tasks = this.tasks.filter(task => task.id !== found.id)
+    // }
 
-    updateTaskStatus(id: string, status: TaskStatus): Task{
-        const task = this.getTaskById(id);
-        task.status = status;
-        return task;
-    }
+    // updateTaskStatus(id: string, status: TaskStatus): Task{
+    //     const task = this.getTaskById(id);
+    //     task.status = status;
+    //     return task;
+    // }
     
 }
